Add shared wallet utilities
