+++ controllers/stationsController.js
@@ -1,4 +1,5 @@
 const db = require('../db');
+const { findEquipmentCostByName, requireFunds, adjustBalance, getBalance } = require('../wallet');
 
 // GET /api/stations
 function getStations(req, res) {
@@ -56,10 +57,47 @@ function deleteStations(req, res) {
   });
 }
 
+// POST /api/stations/:id/equipment  { name: <string> }
+function buyEquipment(req, res) {
+  const stationId = Number(req.params.id);
+  const name = String(req.body?.name || '').trim();
+  if (!stationId || !name) return res.status(400).json({ error: 'Invalid station id or name' });
+
+  db.get(`SELECT equipment, equipment_slots FROM stations WHERE id=?`, [stationId], async (err, st) => {
+    if (err) return res.status(500).json({ error: err.message });
+    if (!st) return res.status(404).json({ error: 'Station not found' });
+
+    let list; try { list = JSON.parse(st.equipment || '[]'); } catch { list = []; }
+    const slots = Number(st.equipment_slots || 0);
+    if (slots && list.length >= slots) return res.status(409).json({ error: 'No free equipment slots' });
+
+    const cost = findEquipmentCostByName(name);
+    try {
+      const ok = await requireFunds(cost);
+      if (!ok.ok) return res.status(409).json({ error: 'Insufficient funds', balance: ok.balance, needed: cost });
+
+      list.push(name);
+      db.run(`UPDATE stations SET equipment=? WHERE id=?`, [JSON.stringify(list), stationId], async (e2) => {
+        if (e2) return res.status(500).json({ error: e2.message });
+        try {
+          await adjustBalance(-cost);
+          const balance = await getBalance();
+          res.json({ success: true, station_id: stationId, equipment: list, cost, balance });
+        } catch (e3) {
+          res.status(500).json({ error: e3.message });
+        }
+      });
+    } catch (e) {
+      res.status(500).json({ error: e.message });
+    }
+  });
+}
+
 module.exports = {
   getStations,
   getStation,
   createStation,
   patchBays,
   deleteStations,
+  buyEquipment,
 };
