+++ routes/stations.js
@@ -7,5 +7,6 @@ router.get('/:id', stations.getStation);
 router.post('/', stations.createStation);
 router.patch('/:id/bays', stations.patchBays);
 router.delete('/', stations.deleteStations);
+router.post('/:id/equipment', stations.buyEquipment);
 
 module.exports = router;
